Add toy test programs for gRPC with library compartmentalisation.
The ping-pong service just bounces requests and responses back and
forth.
The red/blue services simulate the separation of an API into two
compartments. One compartment may have access to more privileged
resources and is kept separate from any failure that may occur in
other parts of the API in this manner.