Improve benchmark run script and setup.
Now we can reuse the jail for multiple run-time configurations as
initially envisioned.
c18n and revocation are enabled by using patchelf and elfctl on the
qps worker before running the benchmark.

Add option to clean the grpc-qps package when re-building with
existing jails.

Add check for a NODEBUG kernel.